requirements.txt
# Using the pipelines target for now to get the full-fledged project Sphinx theme
documenteer[pipelines]>=0.3.0,<0.4.0
ltd-conveyor>=0.4.2,<0.5.0
sqlalchemy>1.1
pyyaml
